Feature/bufr ext (#334)
* Added backend parameter to the YAML file.

* Added support for tranformations.

* Added the ability to do offset and scaling transforms to mnemonic data.

* Fixed bug were there was a type mismatch between the eigen array and values being stored.

* Now using eckit exceptions when encountering configuration issues.

* Updated name of main executable to toIoda.x

* Fixed crash when attempting to use a datetime string variable with an inmemory backend.

* switched back to netcdf backend.

* Changed name of exectuable in CMakeLists.txt

* Made backend strings case insensitive.

+++ src/bufr/DataObject/ArrayDataObject.cpp
@@ -22,7 +22,7 @@ namespace Ingester
     {
         static ioda::VariableCreationParameters params = makeCreationParams();
 
-        auto var = obsGroup.vars.createWithScales<float>(name, dimensions, params);
+        auto var = obsGroup.vars.createWithScales<double>(name, dimensions, params);
         var.writeWithEigenRegular(eigArray_);
         return var;
     }
@@ -37,7 +37,7 @@ namespace Ingester
         ioda::VariableCreationParameters params;
         params.chunk = true;
         params.compressWithGZIP();
-        params.setFillValue<float>(-999);
+        params.setFillValue<double>(-999);
 
         return params;
     }
